requirements-llm.txt
# LLM API clients
openai>=1.0.0
anthropic>=0.8.0

# Text processing
tiktoken>=0.5.0
markdownify>=1.1.0

# HTTP requests
requests>=2.31.0

# Document processing
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 